Jira id: HCLANG-76 - #comment type checking works now
